+++ tests/tracing/conftest.py
@@ -1,8 +1,11 @@
+from unittest import mock
+
 import pytest
 from opentelemetry.trace import _TRACER_PROVIDER_SET_ONCE
 
 import mlflow
 from mlflow.entities import Trace, TraceData, TraceInfo, TraceStatus
+from mlflow.entities.span_status import SpanStatus
 from mlflow.tracing.clients.local import InMemoryTraceClientWithTracking
 from mlflow.tracing.display import IPythonTraceDisplayHandler
 from mlflow.tracing.provider import _TRACER_PROVIDER_INITIALIZED, _setup_tracer_provider
@@ -65,3 +68,26 @@ def create_trace():
 def reset_active_experiment():
     yield
     mlflow.tracking.fluent._active_experiment_id = None
+
+
+@pytest.fixture
+def mock_tracking_serving_client():
+    with mock.patch(
+        "mlflow.tracking._tracking_service.client.TrackingServiceClient.create_trace_info",
+        return_value=TraceInfo(
+            request_id="tr-1234",
+            experiment_id="0",
+            timestamp_ms=0,
+            execution_time_ms=0,
+            status=SpanStatus(TraceStatus.OK),
+            request_metadata={},
+            tags={"mlflow.artifactLocation": "test"},
+        ),
+    ) as mock_create_trace_info, mock.patch(
+        "mlflow.tracking._tracking_service.client.TrackingServiceClient._upload_trace_data",
+        return_value=None,
+    ) as mock_upload_trace_data:
+        yield
+
+        mock_create_trace_info.assert_called()
+        mock_upload_trace_data.assert_called()
